Add undocumented "--interactive" flag
This allows for things like `./example/htop --interactive` (which will launch `htop` in the terminal!) or `./example/htop --interactive sh -c 'ls -l'` which results in `docker run ... --entrypoint sh htop -c 'ls -l'` as one might expect, and appropriately turns `./some-hocker foo bar` into an error (as it should've been previously).

Further room for improvement would be allowing one to modify/override the container name (but share the rest of the arguments) for things like `./example/syncthing --name stcli --interactive syncthing cli show config-status` to be able to run a second container with almost the same arguments without stopping/restarting the original, but I'm less convinced this one's a good idea (which is why this doesn't include it yet).